Crate rust_gpu_tools

Abstraction layer for OpenCL and CUDA.

Feature flags

There are two feature flags, cuda and opencl. By default opencl is enabled. You can enable both at the same time. At least one of them needs to be enabled at any time.

Modules

cuda

The CUDA specific implementation of a Buffer, Device, Program and Kernel.

opencl

The OpenCL specific implementation of a Buffer, Device, Program and Kernel.

Macros

program_closures

Creates two closures, one for CUDA, one for OpenCL for the given one.

Structs

Device

A device that may have a CUDA and/or OpenCL GPU associated with it.

DeviceUuid

A unique identifier based on UUID of the device.

LocalBuffer

A buffer on the GPU.

PciId

The PCI-ID is the combination of the PCI Bus ID and PCI Device ID.

Enums

Framework

Which framework to use, CUDA or OpenCL.

GPUError

Error types of this library.

Program

Abstraction for running programs on CUDA or OpenCL.

UniqueId

Unique identifier that can either be a PCI ID or a UUID.

Vendor

Currently supported vendors of this library.